Improve macOS performance and runtime structure
- Reduce idle polling and snapshot overhead
- Optimize overview previews, animations, and reconciliation
- Expand performance diagnostics and coverage
